Update tests for A1GymEnv and locomotion feature encoder
