+++ tests/test_a1_gym_env.py
@@ -4,8 +4,6 @@ import gym
 import numpy as np
 import torch
 from blind_walking.envs.gym_envs.a1_gym_env import A1GymEnv
-from blind_walking.net.feature_encoder import LocomotionFeatureEncoder
-from gym import spaces
 
 
 class TestA1GymEnv(unittest.TestCase):
@@ -24,13 +22,6 @@ class TestA1GymEnv(unittest.TestCase):
         Reference: line 233 of https://github.com/openai/gym/blob/master/gym/core.py
         """
         return self.env.robot
-
-    def test_default_env_parameters(self):
-        self.env.reset()
-        assert np.all(self.robot.GetMotorPositionGains() == 55.0)
-        assert np.all(self.robot.GetMotorVelocityGains() == 0.6)
-        assert np.all(self.robot.GetMotorStrengthRatios() == 1)
-        assert np.all(self.robot.GetFootFriction() == 0.5), f"{self.robot.GetFootFriction()}"
 
     def test_controller_kp_getter_setter(self):
         self.env.reset()
@@ -59,16 +50,3 @@ class TestA1GymEnv(unittest.TestCase):
         constant = 69.420
         self.robot.SetFootFriction(constant)
         assert np.all(self.robot.GetFootFriction() == constant)
-
-
-class TestLocomotionFeatureEncoder(unittest.TestCase):
-    def setUp(self) -> None:
-        self.env = A1GymEnv()
-        self.extractor = LocomotionFeatureEncoder(self.env.observation_space)
-
-    def test_forward(self):
-        obs = self.env.reset()
-        # Observation is a 1-level dictionary of np arrays
-        # Cast to tensor, dtype float32, and add batch dimension
-        obs_tensor = {k: torch.from_numpy(v).to(torch.float32).view(1, -1) for k, v in obs.items()}
-        features = self.extractor(obs_tensor)
